requirements.txt
neo4j==5.22.0
openai==2.3.0
pinecone-client==4.1.0
pyvis==0.3.1
networkx==3.1
tqdm
python-dotenv
sentence-transformers
